[BEAM-3538] Remove (or merge) Java 8 specific tests module into the main one.
