+++ sdks/java/core/src/test/java/org/apache/beam/sdk/transforms/DistinctTest.java
@@ -17,13 +17,21 @@
  */
 package org.apache.beam.sdk.transforms;
 
+import static org.hamcrest.Matchers.contains;
+import static org.hamcrest.Matchers.hasItem;
+import static org.hamcrest.Matchers.not;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
+import com.google.common.collect.HashMultimap;
+import com.google.common.collect.Multimap;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import org.apache.beam.sdk.coders.KvCoder;
 import org.apache.beam.sdk.coders.StringUtf8Coder;
 import org.apache.beam.sdk.coders.VarIntCoder;
@@ -48,6 +56,7 @@ import org.joda.time.Instant;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
+import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
@@ -56,6 +65,9 @@ import org.junit.runners.JUnit4;
 public class DistinctTest {
 
   @Rule public final TestPipeline p = TestPipeline.create();
+
+  @Rule public ExpectedException thrown = ExpectedException.none();
+
 
   @Test
   @Category(ValidatesRunner.class)
@@ -272,5 +284,54 @@ public class DistinctTest {
 
     PAssert.that(distinctValues).containsInAnyOrder(KV.of(1, "k1"));
     triggeredDistinctRepresentativePipeline.run();
+  }
+
+  @Test
+  public void withLambdaRepresentativeValuesFnNoTypeDescriptorShouldThrow() {
+
+    Multimap<Integer, String> predupedContents = HashMultimap.create();
+    predupedContents.put(3, "foo");
+    predupedContents.put(4, "foos");
+    predupedContents.put(6, "barbaz");
+    predupedContents.put(6, "bazbar");
+    PCollection<String> dupes =
+        p.apply(Create.of("foo", "foos", "barbaz", "barbaz", "bazbar", "foo"));
+
+    thrown.expect(IllegalStateException.class);
+    thrown.expectMessage("Unable to return a default Coder for RemoveRepresentativeDupes");
+
+    // Thrown when applying a transform to the internal WithKeys that withRepresentativeValueFn is
+    // implemented with
+    dupes.apply("RemoveRepresentativeDupes", Distinct.withRepresentativeValueFn(String::length));
+  }
+
+  @Test
+  @Category(NeedsRunner.class)
+  public void withLambdaRepresentativeValuesFnAndTypeDescriptorShouldApplyFn() {
+
+    PCollection<String> dupes =
+        p.apply(Create.of("foo", "foos", "barbaz", "barbaz", "bazbar", "foo"));
+    PCollection<String> deduped =
+        dupes.apply(
+            Distinct.withRepresentativeValueFn(String::length)
+                .withRepresentativeType(TypeDescriptor.of(Integer.class)));
+
+    PAssert.that(deduped).satisfies((Iterable<String> strs) -> {
+      Multimap<Integer, String> predupedContents = HashMultimap.create();
+      predupedContents.put(3, "foo");
+      predupedContents.put(4, "foos");
+      predupedContents.put(6, "barbaz");
+      predupedContents.put(6, "bazbar");
+
+      Set<Integer> seenLengths = new HashSet<>();
+      for (String s : strs) {
+        assertThat(predupedContents.values(), hasItem(s));
+        assertThat(seenLengths, not(contains(s.length())));
+        seenLengths.add(s.length());
+      }
+      return null;
+    });
+
+    p.run();
   }
 }
