+++ sdks/java/core/src/test/java/org/apache/beam/sdk/transforms/SimpleFunctionTest.java
@@ -17,6 +17,10 @@
  */
 package org.apache.beam.sdk.transforms;
 
+import static org.hamcrest.Matchers.equalTo;
+import static org.junit.Assert.assertThat;
+
+import org.apache.beam.sdk.values.TypeDescriptors;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -39,5 +43,34 @@ public class SimpleFunctionTest {
     thrown.expectMessage("apply");
 
     SimpleFunction<Integer, Integer> broken = new SimpleFunction<Integer, Integer>() {};
+  }
+
+  /**
+   * Basic test of {@link MapElements} with a lambda (which is instantiated as a {@link
+   * SerializableFunction}).
+   */
+  @Test
+  public void testGoodTypeForLambda() throws Exception {
+    SimpleFunction<Integer, String> fn = new SimpleFunction<Integer, String>(Object::toString) {};
+
+    assertThat(fn.getInputTypeDescriptor(), equalTo(TypeDescriptors.integers()));
+    assertThat(fn.getOutputTypeDescriptor(), equalTo(TypeDescriptors.strings()));
+  }
+
+  /**
+   * Basic test of {@link MapElements} with a lambda wrapped into a {@link SimpleFunction} to
+   * remember its type.
+   */
+  @Test
+  public void testGoodTypeForMethodRef() throws Exception {
+    SimpleFunction<Integer, String> fn =
+        new SimpleFunction<Integer, String>(SimpleFunctionTest::toStringThisThing) {};
+
+    assertThat(fn.getInputTypeDescriptor(), equalTo(TypeDescriptors.integers()));
+    assertThat(fn.getOutputTypeDescriptor(), equalTo(TypeDescriptors.strings()));
+  }
+
+  private static String toStringThisThing(Integer i) {
+    return i.toString();
   }
 }
