+++ sdks/java/core/src/test/java/org/apache/beam/sdk/transforms/WithKeysTest.java
@@ -25,12 +25,14 @@ import org.apache.beam.sdk.coders.StringUtf8Coder;
 import org.apache.beam.sdk.testing.NeedsRunner;
 import org.apache.beam.sdk.testing.PAssert;
 import org.apache.beam.sdk.testing.TestPipeline;
+import org.apache.beam.sdk.testing.ValidatesRunner;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.TypeDescriptor;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
+import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
@@ -73,6 +75,9 @@ public class WithKeysTest {
 
   @Rule
   public final TestPipeline p = TestPipeline.create();
+
+  @Rule
+  public ExpectedException thrown = ExpectedException.none();
 
   @Test
   @Category(NeedsRunner.class)
@@ -149,5 +154,34 @@ public class WithKeysTest {
     public Integer apply(String value) {
       return value.length();
     }
+  }
+
+  @Test
+  @Category(ValidatesRunner.class)
+  public void withLambdaAndTypeDescriptorShouldSucceed() {
+
+    PCollection<String> values = p.apply(Create.of("1234", "3210", "0", "-12"));
+    PCollection<KV<Integer, String>> kvs = values.apply(
+        WithKeys.of((SerializableFunction<String, Integer>) Integer::valueOf)
+            .withKeyType(TypeDescriptor.of(Integer.class)));
+
+    PAssert.that(kvs).containsInAnyOrder(
+        KV.of(1234, "1234"), KV.of(0, "0"), KV.of(-12, "-12"), KV.of(3210, "3210"));
+
+    p.run();
+  }
+
+  @Test
+  @Category(NeedsRunner.class)
+  public void withLambdaAndNoTypeDescriptorShouldThrow() {
+
+    PCollection<String> values = p.apply(Create.of("1234", "3210", "0", "-12"));
+
+    values.apply("ApplyKeysWithWithKeys", WithKeys.of(Integer::valueOf));
+
+    thrown.expect(IllegalStateException.class);
+    thrown.expectMessage("Unable to return a default Coder for ApplyKeysWithWithKeys");
+
+    p.run();
   }
 }
