+++ settings.gradle
@@ -74,7 +74,6 @@ include ":sdks:java:io:redis"
 include ":sdks:java:io:solr"
 include ":sdks:java:io:tika"
 include ":sdks:java:io:xml"
-include ":sdks:java:java8tests"
 include ":sdks:java:maven-archetypes:examples"
 include ":sdks:java:maven-archetypes:starter"
 include ":sdks:java:nexmark"
